api/requirements.txt
# Essential dependencies for Vercel deployment
fastapi==0.104.1
python-multipart==0.0.6
python-dotenv==1.0.0
google-generativeai==0.3.2
gtts==2.5.1
Pillow==10.1.0
aiofiles==23.2.1
httpx==0.25.2
starlette==0.27.0
pydantic==2.5.0